// src/libs/api/schema/inventory.d.ts
export interface components {
    schemas: {
// ...
        ProductCreate: {
            /** Name */
            name: string;
            /** Price */
            price: number | string | null;
            product_type: components["schemas"]["ProductType"];
            unit: components["schemas"]["Unit"];
            /** Currency Id */
            currency_id: number;
            /** Product Category Id */
            product_category_id: number;
            /** Income Account Id */
            income_account_id: number | null;
            /** Expense Account Id */
            expense_account_id: number | null;
            /** Tags */
            tags: number[];
        };
// ...
        ProductType: "Goods" | "Service";
// ...
        Unit: "Piece" | "Kg" | "Litre" | "Meter" | "Square Meter" | "Cubic Meter" | "Box" | "Bag" | "Bottle" | "Can" | "Tube" | "Roll" | "Set" | "Pair" | "Dozen" | "Pack" | "Ream" | "Bundle" | "Kit" | "Unit" | "Other" | "Unknown" | "Custom";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}